requirements.txt
Flask==2.1.0
Flask-Cors==3.0.10
gunicorn==20.1.0
replicate===0.4.0
urllib3==1.26.12
pytz==2022.2.1
pyairtable==1.4.0
google-api-core==2.10.1
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-tasks==2.10.4
minecraft-launcher-lib==5.3
python-dateutil==2.8.2
